Added converted the format to float, implemented short recording

+++ app/src/main/java/com/exaltead/sceneclassifier/data_extraction/IAudioBufferer.kt
@@ -2,5 +2,5 @@ package com.exaltead.sceneclassifier.data_extraction
 
 interface IAudioBufferer {
 
-    fun takeShortAudioRecord(duration: Double): ShortArray
+    fun takeShortAudioRecord(duration: Double): FloatArray
 }
